requirements.txt
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
et-xmlfile==1.1.0
google==3.0.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
httplib2==0.22.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4.1
tzdata==2023.3
urllib3==1.26.16
